feat: :zap: Improved code quality and maintainability by separating the database connection logic from the server instance.
